refactor(#3683 S2): move the typed-`this` emitters into the subsystem module
The LOC ratchet (#3102) flagged all four lowering-branch call sites — the
god-files have zero headroom. The gate's own remedy ("add code to the subsystem
module, not the barrel/driver") is also the right architecture here, so the
emitter BODIES move into `src/codegen/typed-this.ts` next to the admission
analysis and field resolution, leaving one guarded call plus its rationale
comment at each site:

  tryEmitTypedThisFieldSet   (assignment.ts:        +57 -> +9)
  tryEmitTypedThisCompound   (operator-assignment.ts: +54 -> +9)
  tryEmitTypedThisIncDec     (unary-updates.ts:     +36 -> +10)
  tryEmitTypedThisFieldGet   (property-access-dispatch.ts: +14, already a call)

`EMIT_COMPOUND_OP_HANDLES` moves too — its only consumer is the compound
branch, and its doc now spells out the lockstep coupling with
`emitCompoundOp`'s default-less switch.

Two helpers stay INJECTED rather than imported, to keep the dependency
direction clean: `ensureI32Condition` lives in `codegen/index.ts` and
`emitCompoundOp` is private to `operator-assignment.ts`, both of which sit
above this module. The expression/coercion engines come from `shared.js`, which
exists precisely to offer those late-bound delegates without a cycle.

Net remaining god-file growth is 42 lines across the four files, granted via
`loc-budget-allow` in this issue's frontmatter (down from 161). `check-loc-budget`
now reports OK.

No behaviour change: tsc clean; issue-3683-typed-this-twin (12),
issue-3683-proto-method-write-once, issue-3673-i31-smallint, issue-2151-nary,
issue-2674-* — 42 tests green.

Pre-commit checklist complete ✓
